+++ questionservice/src/main/java/com/quizapp/questionservice/repository/QuestionRepository.java
@@ -14,7 +14,7 @@ public interface QuestionRepository extends JpaRepository<QuestionEntity, Long> 
     List<QuestionEntity> findByCategory(String category);
 
     @Query(value = "SELECT q.id FROM tbl_question q WHERE category =:category ORDER BY RANDOM() LIMIT :num", nativeQuery = true)
-    List<Integer> findRandomQuestionByCategory(String category, int num);
+    List<Long> findRandomQuestionByCategory(String category, int num);
 
      @Query("SELECT new com.quizapp.questionservice.model.QuestionWrapper(q.id, q.questionTitle, q.option1, q.option2, q.option3, q.option4) FROM tbl_question q WHERE q.id IN :ids")
     List<QuestionWrapper> findQuestionsByIds(@Param("ids") List<Long> ids);
